Fix IsNormal, uprank FittingSubgroup method(#61)
Our IsNormal method was buggy (it incorrectly used `Pcp(H,U)`. After fixing this it is mostly equivalent to GAP's default method; so we remove ours and use the default.

This in turn revealed that our `FittingSubgroup` method may not be invoked for finite pcp groups, depending on which packages are loaded. Fix that by increasing its rank.